Refatoração em componentes e styles
