+++ front-end/src/App.jsx
@@ -1,5 +1,5 @@
-import { LoginPage } from "./pages/Login/login"
-import { RegisterPage } from "./pages/cadastro/cadastro";
+import { LoginPage } from "./pages/login/login"
+import { RegisterPage } from "./pages/register/register";
 import "./styles/index.scss";
 
 function App() {
@@ -9,6 +9,7 @@ function App() {
       <div className='App'>
        {/* <LoginPage /> */}
        <RegisterPage />
+       
       </div>
     </>
   )
