+++ front-end/src/components/forms/loginForm/style.module.scss
@@ -1,0 +1,42 @@
+.form{
+    display: flex;
+    flex-direction: column;
+    gap: 1.5625rem;
+    margin-top: 1.875rem;
+    background: white;
+    padding: 40px;
+    border-radius: 4px;
+
+    & > div:first-child{
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        flex-direction: column;
+        gap: 20px;
+        h2 {
+            color: #000;
+            font-size: 1rem;
+            
+        }
+    }
+    
+    & > div:last-child{
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        gap: 1.25rem
+    }
+    .footer{
+        span {
+            font-family: var(--font-primary);
+            font-size: 1rem;
+            font-weight: 300;
+        }
+        p {
+            font-family: var(--font-primary);
+            color: var(--color-secondary);
+            font-size: 1rem;
+            font-weight: 700;
+        }
+    }
+}
