+++ front-end/src/components/forms/registerForm/formSchemaRegister.js
