+++ front-end/src/components/forms/registerForm/style.module.scss
@@ -1,0 +1,23 @@
+.form{
+    display: flex;
+    flex-direction: column;
+    gap: 1.5625rem;
+    margin-top: 1.875rem;
+    background: white;
+    padding: 42px 22px;
+    border-radius: 4px;
+
+    & > div:first-child{
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        gap: 15px;
+    }
+    
+    & > div:last-child{
+        display: flex;
+        flex-direction: column;
+        gap: 1.25rem;
+    }
+}
