+++ front-end/src/pages/login/login.jsx
@@ -1,0 +1,17 @@
+import styles from "./style.module.scss";
+import Vector from "../../assets/Vector.png";
+import { LoginForm } from "../../components/forms/loginForm/loginForm";
+
+export const LoginPage = () => {
+
+  return (
+    <main>
+      <div className="container">
+        <div className={styles.flexBox}>
+          <img src={Vector} alt="Vector image" />
+          <LoginForm />
+        </div>
+      </div>
+    </main>
+  );
+};
