+++ front-end/src/pages/register/style.module.scss
@@ -1,0 +1,37 @@
+.flexBox {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 0.625rem;
+    margin-top: 50px;
+  
+    img{
+      height: 1.3257rem;
+    }
+  
+    & > div:first-child{
+      width: 100%;
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+    }
+
+  @media (min-width:320px) {
+    label{
+      margin-left: 20px;
+    }
+    input{
+      width: 90%;
+      margin-left: 20px;
+    }
+    select{
+      width: 90%;
+      margin-left: 20px;
+    }
+  }
+   
+    button{
+      width: 90%;
+      margin-left: 20px;
+    }
+  }
