+++ front-end/src/styles/global.scss
@@ -1,0 +1,6 @@
+:root{
+    --font-primary:'Inter', sans-serif;
+    --color-primary: #4A2A50;
+    --color-secondary: #82498C;
+    --color-primary-focus: #FF427F;
+}
