+++ front-end/src/styles/grid.scss
@@ -2,8 +2,5 @@
     width: 100%;
     margin: 0 auto;
     padding: 1rem 0;
-
-    &.sm{
-        max-width: 400px;
-    }
+    max-width: 1200px;
 }
