+++ front-end/src/styles/typography.scss
@@ -1,0 +1,18 @@
+.title{
+    font-family: var(--font-primary);
+    font-weight: 700;
+
+    &.one{
+        font-size: clamp(2rem, 5vw, 3.2rem);
+    }
+
+    &.two{
+        font-size: clamp(1.5rem, 4vw, 2.8rem);
+    }
+}
+
+.paragraph{
+    font-family: var(--font-primary);
+    font-weight: 400;
+    font-size: 1rem;
+}
