--- src/pages/Login/style.module.scss
@@ -1,54 +1,0 @@
-main {
-    display: flex;
-    align-items: center;
-}
-.flexBox{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: .625rem;
-    margin-top: 50px;
-
-    form{
-        display: flex;
-        flex-direction: column;
-        gap: 1.5625rem;
-        margin-top: 1.875rem;
-        background: white;
-        padding: 42px 22px;
-        border-radius: 4px;
-
-        & > div:first-child{
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            flex-direction: column;
-            gap: 20px;
-            h2 {
-                color: #000;
-                font-size: 1rem;
-                
-            }
-        }
-        
-        & > div:last-child{
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            gap: 1.25rem
-        }
-        .footer{
-            span {
-                font-family: var(--font-primary);
-                font-size: 1rem;
-                font-weight: 300;
-            }
-            p {
-                font-family: var(--font-primary);
-                color: #82498C;
-                font-size: 1rem;
-                font-weight: 700;
-            }
-        }
-    }
-}
