--- src/pages/cadastro/style.module.scss
@@ -1,70 +1,0 @@
-.flexBox {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0.625rem;
-    margin-top: 50px;
-  
-    img{
-      height: 1.3257rem;
-    }
-  
-    & > div:first-child{
-      width: 100%;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-    }
-    form{
-      display: flex;
-      flex-direction: column;
-      gap: 1.5625rem;
-      margin-top: 1.875rem;
-      background: white;
-      padding: 42px 22px;
-      border-radius: 4px;
-  
-      & > div:first-child{
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          justify-content: center;
-          gap: 15px;
-      }
-      
-      & > div:last-child{
-          display: flex;
-          flex-direction: column;
-          gap: 1.25rem;
-      }
-      select{
-        height: 48px;
-        background: var(--color-grey2);
-        color: #868E96;
-        font-size: 16px;
-        font-family: var(--font-primary);
-        font-weight: 400;
-        border-radius: 4px;
-        display: flex;
-        align-items: center;
-      }
-  }
-  @media (min-width:320px) {
-    label{
-      margin-left: 20px;
-    }
-    input{
-      width: 90%;
-      margin-left: 20px;
-    }
-    select{
-      width: 90%;
-      margin-left: 20px;
-    }
-  }
-   
-    button{
-      width: 90%;
-      margin-left: 20px;
-    }
-  }
