--- src/styles/global.scss
@@ -1,11 +1,0 @@
-:root{
-    --font-primary:'Inter', sans-serif;
-
-    --color-primary: #FF577F;
-    --color-primary-focus: #FF427F;
-    --color-grey4: #121214;
-    --color-purple: #4A2A50;
-    --color-grey2: #343B41;
-    --color-grey1: #868E96;
-    --color-grey0: #F8F9FA;
-}
